requirements.txt
# Core ML & Data Science
numpy>=2.3,<2.4
pandas>=2.0,<2.1
scikit-learn>=1.5.2
xgboost>=2.0,<2.1
matplotlib>=3.7,<3.8
plotly>=5.17,<6.0

# SHAP для интерпретации
shap>=0.44,<0.45
numba>=0.62,<0.63
tqdm>=4.67,<4.70
slicer>=0.0.7,<0.1
cloudpickle>=3.1,<3.2

# Web Framework
flask>=2.3,<2.4
flask-cors>=4.0,<5.0
Werkzeug>=3.1.3,<4.0
Jinja2>=3.1.6,<4.0
itsdangerous>=2.2,<3.0
click>=8.3,<9.0
blinker>=1.9,<2.0

# Database
psycopg2-binary>=2.9,<3.0

# Dashboard
streamlit>=1.29,<2.0
altair>=5.5,<6.0
cachetools>=5.5,<6.0
importlib-metadata>=6.11,<7.0

# OpenAI (опционально для AI рекомендаций)
openai>=0.28,<0.29

# Дополнительные утилиты
requests>=2.31,<3.0
python-dateutil>=2.8,<3.0
pytz>=2023.3,<2024.0
